Image / File display in edit form
If there is a collection of file field inside the form (for a custom resource with a collection of images for example) and there one or more image are persisted, display a thumbnail of that image for each. Currently there is nothing displayed, the file field in the edit page looks empty even if there is already a file saved, so it's a bit confusing for the admin.